refactor(lib): content block extension
